공통 컴포넌트 수정 및 추가: 기존 콤보박스 수정 및 새로운 콤보박스 컴포넌트 생성
